+++ L14_1-threads/pom.xml
@@ -8,5 +8,13 @@
     <artifactId>L14_1-threads</artifactId>
     <version>1.0-SNAPSHOT</version>
 
+    <dependencies>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.12</version>
+        </dependency>
+    </dependencies>
+
     
 </project>
